requirements.txt
python-dotenv==1.2.1
prettytable==3.9.0
python-telegram-bot==22.5
